authentication added for all the operations

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is :", req.body);
   const { name, email, phone } = req.body;
@@ -23,13 +23,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 //@desc Get contacts
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -41,12 +42,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update contact
 //@route POST /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    req.status(403);
+    throw new Error("user don't have permission to update other user contact");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -59,12 +65,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    req.status(403);
+    throw new Error("user don't have permission to delete other user contact");
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
